refactor: moves files to the better package
